fix: correct backup functions

+++ private-files/helper-functions.sh
@@ -2,16 +2,20 @@
 # backup and restore scripts.
 
 # Helper function to backup files after checking with the user first.
-function backup_safe() {
-    echo -n "Backup '$1' to '$2'? [y/n]: "
-    read yesno
-    if [[ $yesno =~ ^[Yy] ]]; then
-        aws s3 cp "$1" "$2" $3 --profile "${profile}"
+backup_safe() {
+    if [ ! -r "${1}" ]; then
+        echo "skipping missing file '${1}'..."
+    else
+        echo -n "backup '$1' to '$2'? [y/n]: "
+        read yesno
+        if [[ $yesno =~ ^[Yy] ]]; then
+            aws s3 cp "$1" "$2" $3 --profile "${profile}"
+        fi
     fi
 }
 
 # Helper function to restore files after checking with the user first.
-function restore_safe() {
+restore_safe() {
     echo -n "Restore '$1' to '$2'? [y/n]: "
     read yesno
     if [[ $yesno =~ ^[Yy] ]]; then
@@ -24,8 +28,7 @@ function restore_safe() {
 # Helper function to check if an AWS profile exists.
 # Returns: 0 if the profile exists, 1 if it doesn't and 2 if the config doesn't
 # exist.
-function aws_profile_exists() {
-
+aws_profile_exists() {
     local profile_name="${1}"
     local config_path="${HOME}/.aws/config"
     echo -n "Checking for AWS profile: '${profile_name}'... "
